+++ tests/meltano/core/plugin/singer/test_tap.py
@@ -754,6 +754,49 @@ class TestSingerTap:
         assert not catalog_path.exists(), "Catalog should not be present."
 
     @pytest.mark.asyncio
+    async def test_run_discovery_stderr_output(
+        self,
+        plugin_invoker_factory,
+        session,
+        subject,
+        elt_context_builder,
+        project_plugins_service,
+    ):
+
+        process_mock = mock.Mock()
+        process_mock.name = subject.name
+        # we need to exit successfully to not trigger error handling
+        process_mock.wait = CoroutineMock(return_value=0)
+        process_mock.returncode = 0
+        process_mock.stderr.at_eof.side_effect = (False, True)
+        process_mock.stderr.readline = CoroutineMock(return_value=b"stderr mock output")
+        process_mock.stdout.at_eof.side_effect = (True, True)
+        process_mock.stdout.readline = CoroutineMock(return_value=b"")
+
+        invoker = plugin_invoker_factory(subject)
+        invoker.invoke_async = CoroutineMock(return_value=process_mock)
+        catalog_path = invoker.files["catalog"]
+
+        with mock.patch(
+            "meltano.core.plugin.singer.tap._stream_redirect"
+        ) as stream_mock:
+            with mock.patch(
+                "meltano.core.plugin.singer.tap.logger.isEnabledFor", return_value=False
+            ):
+                await subject.run_discovery(invoker, catalog_path)
+                assert stream_mock.call_count == 1
+
+        with mock.patch(
+            "meltano.core.plugin.singer.tap._stream_redirect"
+        ) as stream_mock:
+            with mock.patch(
+                "meltano.core.plugin.singer.tap.logger.isEnabledFor", return_value=True
+            ):
+                await subject.run_discovery(invoker, catalog_path)
+                assert stream_mock.call_count == 2
+                assert stream_mock.call_args[1]["write_str"] is True
+
+    @pytest.mark.asyncio
     async def test_run_discovery_handle_io_exceptions(
         self,
         plugin_invoker_factory,